feat : add logout service
