--- src/app/auth/signup/index.interface.ts
@@ -1,19 +1,0 @@
-export interface ISubmitSignupForm {
-  username: string
-  password: string
-  name: string
-  email: string
-  // remember?: boolean
-}
-
-export interface IErrorResponse {
-  code: number
-  message: string
-}
-
-export interface IResponseDataSignup {
-  accessToken: string
-  email: string
-  name: string
-  role: string
-}
