--- src/components/organisms/Auth/LoginForm/index.interface.ts
@@ -1,17 +1,0 @@
-export interface ISubmitLoginForm {
-  username: string
-  password: string
-  remember?: boolean
-}
-
-export interface IErrorResponse {
-  code: number
-  message: string
-}
-
-export interface IResponseDataLogin {
-  accessToken: string
-  email: string
-  name: string
-  role: string
-}
