--- src/components/organisms/Auth/SignupForm/signup.module.scss
@@ -1,44 +1,0 @@
-.card {
-  box-shadow: 0px 3px 5px 0px rgba(0, 0, 0, 0.1);
-  filter: none;
-  z-index: 2;
-}
-
-.signup-container {
-  position: relative;
-  @apply flex w-full items-center justify-center bg-white p-10 pt-40;
-
-  &::before {
-    content: '';
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: 0;
-    background-image: url('https://pce.sandiego.edu/wp-content/uploads/2022/11/private-foundation-vs-public-charity.jpg');
-    // background-image: url("path/to/your/image.jpg");
-    background-size: cover;
-    background-position: bottom;
-    filter: blur(1px);
-    z-index: 1;
-  }
-  background-repeat: no-repeat;
-}
-.antd-form-item {
-  label {
-    @apply text-xs font-medium uppercase tracking-wider text-gray-500;
-  }
-  input {
-    @apply mt-1.5 block w-full rounded bg-gray-50 px-4 py-3 tracking-wide text-gray-800 focus:outline-none focus:ring-2 focus:ring-gray-200;
-  }
-}
-
-.checkbox-remember {
-  label {
-    display: flex;
-    align-items: center;
-
-    span:nth-child(2) {
-      white-space: nowrap;
-    }
-  }
-}
